GSOP-236 add constructor with parameters
